feat: add confirmation alert when delete job

+++ vue-project/src/views/JobView.vue
@@ -19,9 +19,13 @@ const state = reactive({
 
 const deleteJob = async (id) => {
   try {
-    await axios.delete(`/api/jobs/${jobId}`);
-    toast.success("Job Deleted Successfully");
-    router.push("/jobs");
+    const confirm = window.confirm("Are you sure you want to delete this job?");
+
+    if (confirm) {
+      await axios.delete(`/api/jobs/${jobId}`);
+      toast.success("Job Deleted Successfully");
+      router.push("/jobs");
+    }
   } catch (error) {
     toast.success("Job Was Not Deleted");
   }
